LayoutTests/imported/w3c/web-platform-tests/css/css-tables/table-cell-writing-mode-computed-expected.txt

Test

FAIL Computed value of orthogonal writing-mode on table cell assert_equals: expected "vertical-lr" but got "horizontal-tb"
